requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
black==19.3b0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
codecov==2.0.15
coloredlogs==9.0
coverage==4.5.3
cryptography==2.3.1
decorator==4.4.0
ecdsa==0.13
ECPy==0.8.3
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
fakeredis==1.0.3
fancycompleter==0.8
Flask==1.0.2
Flask-Cors==3.0.3
future==0.17.1
gunicorn==19.8.1
hidapi==0.7.99.post21
humanfriendly==4.18
idna==2.7
importlib-metadata==0.18
ipaddress==1.0.22
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jsonschema==2.6.0
ledgerblue==0.1.16
lorem==0.1.1
MarkupSafe==1.1.1
more-itertools==7.1.0
mypy==0.711
mypy-extensions==0.4.1
olefile==0.44
packaging==19.0
parso==0.5.0
pdbpp==0.10.0
peewee==3.8.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.12.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.5.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
pudb==2019.1
py==1.8.0
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.8.2
Pygments==2.4.2
pyparsing==2.4.0
pytest==5.0.1
pytest-cov==2.7.1
pytest-sugar==0.9.2
python-dateutil==2.7.3
python-u2flib-host==3.0.3
PyYAML==5.1
redis==2.10.6
requests==2.20.0
secp256k1==0.13.2
singleton-decorator==1.0.0
six==1.12.0
sortedcontainers==2.1.0
termcolor==1.1.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.24.3
urwid==2.0.1
uWSGI==2.0.18
wcwidth==0.1.7
Werkzeug==0.14.1
wmctrl==0.3
zipp==0.5.2
